refactor: reduce number of test files
